+++ Cargo.toml
@@ -13,6 +13,7 @@ required-features = ["build-binary"]
 [dependencies]
 anyhow = "1.0.75"
 clap = { version = "4.4.11", features = ["cargo", "derive"], optional = true }
+html-escape = "0.2.13"
 rand = "0.8.5"
 reqwest = { version = "0.11.22", features = ["cookies", "blocking"] }
 serde = { version = "1.0.193", features = ["derive"] }
